Split change-file-owner into set-file-{owner,group}!, add file-group
An accessor for the file's group ownership was missing, so we add
this.

Just like for permissions, it makes more sense to have a symmetric set
of getter & setter procedures, with a matching SRFI-17 setter, so the
setter is renamed and split in two.  This is also moved from
(chicken file) to (chicken file posix), where it belongs.

We also use fchmod() if a port or FD is passed in to make it
consistent with the getter.

Signed-off-by: Kooda <kooda@upyum.com>