Replace convert_string_to_number with simpler flonum-only decoder
Now that decode_literal() no longer needs to be able to deal with the
case that compile-time fixnums might decode at runtime to flonums when
going from a 64-bit machine to a 32-bit machine, we can simplify this
code a lot.

We now also get rid of the "radix" parameter, because string->number
no longer relies on convert_string_to_number.  We could have removed
"radix" sooner, but it made more sense to postpone until now.

Similarly, because we only ever need to decode literals produced by
the compiler, rather than arbitrary user input, we can radically
simplify the parsing: just call strtod() and be done with it.  We only
need to take care about parsing nan and inf syntax, because ours might
not coincide with libc's.

The decode_literal() code itself is also slightly restructured:
originally, a "large fixnum" would be encoded as \xff\xfe...\0 and a
proper flonum would be encoded as just \xfe...\0.  Now, we only need
to process the latter encoding.  Unfortunately, because size is not
encoded in the string, we have to special-case the flonum type a
little bit: instead of adding it to the final switch(), we have to do
an if() check before the size is decoded.

Signed-off-by: Evan Hanson <evhan@foldling.org>