Ensure some functions marked OBSOLETE are really unused
These were supposed to have been removed before, but they were still
used in the compiler.

C_u_i_positivep and C_u_i_evenp, as well as C_u_i_evenp and C_u_i_oddp
are supposed to be unsafe faster variants of their C_i_-prefixed
brethren, but because the type analysis is more complicated now that
we have more than 2 numeric types, the additional error check is not
worthwhile to leave off.

C_i_foreign_integer_argumentp and ##sys#foreign-integer-argument
(which is its Scheme procedure) were still used by the "enum" foreign
type specifier.  This has now been replaced by a ranged integer check.
According to the C spec, enums are represented exactly as integers, so
we just use the same range as integers.  The test suite was lacking a
test for the enum foreign type, so this patch adds that as well.

Finally, the rewrites for C_i_zerop and C_u_i_zerop were the wrong way
around.  Our rewrites are applied in order, and a safe rewrite is
always allowed, so it would never get to the unsafe rewrite.  We fix
this by swapping them.  Now, the unsafe rewrite is considered first,
and if we're in safe mode, we reject it and try the safe one instead.

Signed-off-by: Evan Hanson <evhan@foldling.org>