Add a more convenient way of constructing condition objects
This "condition" constructor is inspired by Kon Lovett's
"make-condition+" constructor from the condition-utils egg.

This also adds a helper procedure for converting plist-style condition
property lists to the internal structure of condition properties by
consing the kind onto the property, followed by the value.  This also
is used in make-property-condition, which now gives a better error
message when the property list argument isn't a list with an even
element count.

Signed-off-by: Evan Hanson <evhan@foldling.org>