In the scrutinizer, do not assume big fixnums will fit into 32 bits
When the scrutinizer applies a specialization for a fixnum, it should
make sure it really is a fixnum. If compiling on a 64-bit platform, a
fixnum literal might no longer be a fixnum when the program is running
on a 32-bit platform.

Thus, we check whether the literal is a big-fixnum? first.

Similarly, small bignums are rewritten to 'integer for the reverse
situation: when compiling on 32-bit, a bignum might become a fixnum when
running on 64-bit.

Signed-off-by: Evan Hanson <evhan@foldling.org>