In the scrutinizer, do not assume big fixnums will fit into 32 bits
When the scrutinizer applies a specialization for a fixnum, it should
make sure it really is a fixnum. If compiling on a 64-bit platform, a
fixnum literal might no longer be a fixnum when the program is running
on a 32-bit platform.

Thus, we check whether the literal is a big-fixnum? first.

Unfortunately, we can't do the converse in CHICKEN 4: when compiling on
a 32-bit platform, a "big fixnum" literal will always be read in as a
flonum. However, this will result in consistent behaviour when
compiling the resulting C code on 32-bit or 64-bit.

Signed-off-by: Evan Hanson <evhan@foldling.org>