Move `module-environment' into chicken.eval
To align more closely with R7RS's library organisation, we put the
`module-environment' procedure in the "chicken.eval" module.

Also, drop the [undocumented] second argument from the user-facing
version of this procedure, and remove it from the plain "chicken"
module. The dropped argument is still used internally, but doesn't need
to be part of the public API.

Signed-off-by: Peter Bex <peter@more-magic.net>