Rename (chicken expand) to (chicken syntax)
This adds a small hack to the chicken.import.scm import library to
define chicken.syntax#er-macro-transformer and friends as aliases for
their chicken.expand# version, if those are defined.

This is required when building the new chicken with an old one,
because the old compiler will load "chicken.import.scm", but the name
mappings aren't right for anything that happens at expansion time if
there are no definition from the chicken.syntax module inside that
compiler's libchicken.

We also make some internal definitions in expand.scm explicit, namely
match-expression and expand-curried-define.  This happened to work
because the macros that use these definitions live inside the same
unit of compilation.  If we were to move them, things would
mysteriously start breaking.

Signed-off-by: Evan Hanson <evhan@foldling.org>