Export CHICKEN_REPOSITORY_PATH for functor and linking tests
This is one half of 934eecc re-applied to fix a problem when running
`make check` without first running `make install`: when compiling and
executing use-square-functor.scm in the functor tests, the import
library for the "chicken" module wouldn't be found since it wasn't
present in the target install directory. To fix this, we export the
CHICKEN_REPOSITORY_PATH and CHICKEN_INSTALL_REPOSITORY variables so that
the test executable looks there rather than in the install directory.

Signed-off-by: Peter Bex <peter@more-magic.net>