Finalize (chicken load) by hiding `load' and `load-extension'
`load-extension' was deprecated in favor of `require'. `load' must be
exported by scheme, and it makes no sense to export it in two modules
so we do not export it from (chicken load).

This also adds those type database entries which were missing for the
remaining exported procedures in (chicken load).

Signed-off-by: Evan Hanson <evhan@foldling.org>