Fix unpersistability sanity check for symbol GC
The sanity check we do just before dropping symbols from the symbol
table will ensure that the "bound value" slot of a symbol is either
empty or contains the symbol itself.  If either is violated, the
assertion check fails.

This is correct, but keywords may have been moved by the GC and
replaced with forwarding pointers.  This means the symbol value slot
could contain a forwarding pointer, which is definitely _not_ equal to
the symbol itself.

To fix this, before performing our sanity checks, we now fix up any
forwarding pointers in the symbol's value.

Signed-off-by: felix <felix@call-with-current-continuation.org>