Simplify forwarding pointer wrapping/unwrapping
The original version would take the high bit and put it in the low
bit, clobbering what was there.  That's okay because pointers are
word-aligned so they won't have their lowest bit set anyway.

Then the high bit would be set to mark the header as a forwarding
pointer.  But a much simpler way of achieving the same is to simply
shift right by one and set the high bit.

Signed-off-by: Evan Hanson <evhan@foldling.org>