Get rid of "global namespace" for record types.
Before, when define-record-type or define-record were used, the name
that's passed in is taken directly as the tag. Therefore, if two
completely unrelated modules define a record type (with a completely
different layout), both predicates would accept each other's record
objects. Worse, if an accessor was used on one of them this could be an
unsafe operation resulting in segfaults and other Bad Things.

This restores "applied zbigniew's record-rename patch (finally), added
test-case", which is 727b2b3fea271474540f215af4842d32e82e7e6d. Thereby,
we revert commit 045524a45217ff1eed79a73190d61c561999fba5.

Furthermore, we now expose the "type name" as an identifier which is
bound to the value of the type tag stored in the record. This is a
requirement of both SRFI-99 and R7RS (see #1342), and ensures we can
access the tag of a record type defined in another module. This is
useful for the record-variants egg; it extends existing/predefined
record types with new behaviour, but to do so it needs to be able to
know what the tag is even if the record was defined in another module.

Signed-off-by: Evan Hanson <evhan@foldling.org>