Fix repository searching for "-link" flag and associated tests
The part of the linking tests that was supposed to make sure the "-link"
flag picks up object files from the repository was broken, since the
reverser.o file that should have been *moved* into the test repository
was *copied* there instead. This meant that a reverser.o was left in the
current directory, and the test picked up that one rather than the one
from the repo.

Fixing this uncovered the fact that the behaviour it was meant to test
was broken, too, so this patch also fixes that by updating the
`find-object-file` procedure in csc.scm to use the same lookup logic
that `##sys#process-require` uses for static extensions. It also factors
some duplicated logic for finding files in the repository out into a new
`find-file` helper in the load module, standardises the naming of the
two procedures used to look up files for extensions, and uses internal
namespacing to share these procedures between files.

Also, use "rmdir" to make sure the test repository is always deleted
before running the test suite on Windows.

Signed-off-by: felix <felix@call-with-current-continuation.org>