Add compatibility shims for use of 'define-record-type' within core
After 54b0d5ad, record type tags include a namespace prefix when defined
within a module. However, when bootstrapping from an older version of
CHICKEN, the compiler that's used to build the stage1 boot-chicken won't
include these prefixes and the resulting chicken will bomb out with
errors when building core. As a temporary workaround, we add a feature
as an indicator of the new behaviour and revert to the old behaviour
when it isn't registered.