Rename bit-set? to bit->boolean to avoid confusion (fixes #1385)
The problem with bit-set? is that our definition has the argument
order swapped when compared to SRFI-33 and SRFI-60.  Given that all
our other procedures follow the definitions given in these SRFIs, it
is extra confusing that this one procedure has a different argument
order.  This may result in very subtle bugs.

To make matters worse, swapping the argument to match the SRFIs would
be downright evil, because it would make porting bugs harder to
find: (bit-set? 1 2) for example will return different values
depending on which argument indicates the number and which the bit
position, but the result is still a boolean and in other cases it
might "accidentally" return the expected result, making it very very
difficult to figure out why a program is failing.

So this is why we rename it: When porting any program from CHICKEN 4
to CHICKEN 5 (or from another Scheme), it will immediately error out,
and after a quick search one will be able to find the CHICKEN 5
procedure bit->boolean (and curse us for deviating from the SRFI, not
knowing our alternatives were even worse).

The new bit->boolean procedure immediately has a sort of deprecated
status.  Later on, after enough time has passed to have ported all
CHICKEN 4 code, bit-set? may be re-introduced with the
correct (SRFI-compliant) argument order, and we can then officially
deprecate bit->boolean.  Even later still we can finally get rid of
this ugly temporary procedure.

Signed-off-by: Kooda <kooda@upyum.com>