Move sorting procedures from chicken.data-structures to chicken.sort
This also wraps the chicken-bug and chicken-profile programs in a
"main" module.  This helped find a bug in chicken-profile: we forgot
to import chicken.file in chicken-profile when we moved "glob" into
that module.

We also forgot to export merge!, the side-effect version of merge from
data-structures.  The new chicken.sort module does export this
procedure.

Signed-off-by: Kooda <kooda@upyum.com>