Make `current-directory` a pure getter with associated setter
This adds a SRFI-17 setter for `current-directory` and drops the
optional argument that could previously be passed to cause it to change
directories.

This makes it behave more like the other procedures in the posix unit,
and makes for a better API since the current directory is a
process-level variable and not a SRFI-39 parameter object (and it can't
be made into one, since the CWD is not thread-local).

Signed-off-by: Peter Bex <peter@more-magic.net>