Add "chicken.type" module
This moves the type-related macros in chicken-syntax.scm upwards in the
file and captures a syntax environment containing just those six forms.
These are added to a new "chicken.type" module defined in module.scm. We
also leave these macros in the bare "chicken" module for the time being,
as we have done with the "chicken.condition" module.

Signed-off-by: Peter Bex <peter@more-magic.net>