Add preliminary version of (chicken base) module
This is a primitive module because it has several syntax exports.

We first define the module containing mostly stub definitions which
are set! further down in the file.  This ensures we won't have to
import the module into itself, which will result in strange/unexpected
errors.

There's one hack: at the end of library.scm, we define a few toplevel
aliases for procedures that are supposed to be in the default macro
namespace.  Without this, we'd need (import-for-syntax chicken.base)
in practically every low-level macro use.  We will need to find out
why this is needed, and take better control over what's available in
the initial macro environment.

Signed-off-by: Evan Hanson <evhan@foldling.org>