Add a more thorough workaround for stat() issues on MinGW
MinGW's stat(2) syscall is badly behaved with respect to trailing
slashes in pathnames, so we add a drop-in replacement ("C_stat") in
chicken.h that smoothes over these problems on MinGW, MSYS, and MSYS2.

In particular, this replacement will strip trailing slashes and retry
when the built-in stat(2) yields ENOENT, since paths like "<dir>//" will
cause MinGW's version to fail (incorrectly). It will also return -1 when
given a non-directory pathname that includes a trailing slash, since the
built-in version will actually succeed in this case when it should
instead fail with errno set to ENOTDIR.

Note that we must attempt to stat the given filename at least once
before stripping slashes, since there are some valid paths (e.g. "C:/"
or "\\") that require them to be present in order for the call to
succeed.

The preexisting stat macros in posix*.scm have been given "u" and "i"
prefixes to avoid name collisions with the new wrapper, and to indicate
their proper usage.

This commit also adds a test script to check for these and other similar
problems on all platforms. The invocation of this script in runtests.sh
is a bit unusual, passing "//" as an argument instead of "/" when
running under MSYS*. This is because a single "/" will expand to the
MSYS installation's root directory, for example "C:/MinGW/msys/1.0"; the
second slash effectively escapes the first and disables this behaviour.

Signed-off-by: Peter Bex <peter@more-magic.net>