Threads are tricky business. We must sacrifice a goat in the name of the dark gods to make them work properly!
When scheduling a new thread, we need to invoke it with a continuation
that goes nowhere.  We know the thread call won't return, so the
continuation must be swapped out because the scheduler continuation
still has a reference to the old thread before invocation.  This
continuation itself still may holds a reference to the previous
thread, and so on, resulting in infinite memory buildup.

This fixes #1367

Signed-off-by: Kooda <kooda@upyum.com>