Delete things marked OBSOLETE for the pre2 snapshot
We'll introduce a couple more with the (chicken base) change and
perhaps others, too.  So let's remove the ones which are no longer
necessary.

Note: the scheduler used the flonum 0.0 as a default timeout value as
an argument to a foreign-lambda that accepts an unsigned-int, this had
to be changed to fixnum 0 to avoid a crash due to the removal of the
OBSOLETE compatibility for automatically coercing flonums to these
argument types.  This crash only happens because the scheduler has an
`unsafe' declaration.

Signed-off-by: Evan Hanson <evhan@foldling.org>