Replace all uses of C_mutate2() with C_mutate()
C_mutate2 has been deprecated for a long time (it has been in core
since 6c90c4a1a89 in 2012, as a temporary compatibility workaround),
but calls to it were *still* being generated by the compiler.  In
order to get rid of the definition, we first need to ensure we don't
generate any calls to it ourselves.

This means that after the next snapshot we can finally and completely
remove C_mutate2().

Signed-off-by: Evan Hanson <evhan@foldling.org>