Add unit dependency info to the scheme and rnrs[-null] modules
The r4rs module depends on the "library" unit, so we add that dependency
to its module registration in modules.scm.

The r[45]rs-null modules don't depend on any unit since they only export
syntax, so set their module-library slots to #f to indicate "no unit".

The scheme module only depends on the "library" unit, so remove it from
the list of `core-unit-requirements` and add "library" to its module
registration so that it's treated like any other unit.

Add a test to make sure the r4rs and r4-null libraries are working.

Signed-off-by: Peter Bex <peter@more-magic.net>