Fix flonum to string conversion on extreme edge cases on 64-bit archs
This bug was exposed by the previous commit which restored the tests
to work as intended, because the serialization of (expt 2.0 64)
resulted in "0.0" in compiled code, causing the exact/inexact
comparison tests between the bignum (expt 2 64) and the flonum
(expt 2.0 64.0) to fail.

This representation fails because C_fits_in_unsigned_int_p will
incorrectly return C_SCHEME_TRUE on a 64-bit architecture due to how
we perform the casting: Because UINT_MAX is 2^64-1, it will be
normalized by the floating-point representation to the nearest
floating-point value that can represent it: 2^64.
Thus, (double)2^64 <= UINT_MAX is true.  When we then cast the
double to an unsigned word, it will result in zero due to overflow.

The fix here is to simply check if the log2 of the number is less than
the number of bits in an UWORD, rather than trying a direct value
comparison.  As a bonus, this allows us to finally get rid of the
obsolete C_fits_in_unsigned_int_p() inline helper function.

Signed-off-by: Peter Bex <peter@more-magic.net>