Add initial version of the "scheme" module.
This makes a few things more explicit and even pointed out some
issues:

- "cond-expand" needs to compare unhygienically, because the initial
  environment will contain "and", "or" and "not" from scheme, but
  it should work even if "scheme" is not imported, so we can't
  compare equality like that.  The Chibi implementation (reference
  implementation for R7RS) also does this unhygienically.
- "module" needs to compare "=" (for functor syntax) unhygienically
  for the same basic reason.
- "Extended bindings" included things like exact-integer?, make-polar,
  make-rectangular, real-part, imag-part and string->symbol, which
  are all *standard* bindings, from r5rs.
- There are still a few macros in the wrong place, like delay-force,
  cond-expand, require-library and letrec*
- A few macro definitions were missing some mappings in their
  syntactic environments.

Also, a few definitions from chicken.base were still unprefixed in
some places, those are now prefixed.

Last but not least, in a few places in compiler-syntax, the bindings
checked in standard-bindings were missing their prefix, thus causing
the compiler specialization to not be applied.

The define-internal-compiler-syntax macro has been slightly simplified
by quoting the syntax environments in the call rather than in the
macro itself, and the r-c-s procedure is only used by the macro, so
there's no need to make the syntactic environment argument optional:
we always expand to a call that includes it.

Signed-off-by: felix <felix@call-with-current-continuation.org>
(crosses fingers)