Clean up "scheme" module to only contain standard definitions.
In order to more tightly control what gets included in the
"scheme" (or rnrs[-null]) module, we capture just the standard macro
definitions and bind them to ##sys#scheme-macro-environment, which is
used in the hardcoded rnrs module definitions in modules.scm.

The macros "syntax", "letrec*", "delay-force" and "require-library" are
moved to the (chicken base) module, where they belong. Because letrec*
uses "check-for-multiple-bindings", the definition of this has been
moved to (chicken internal), as well as the "macro-subset" helper
procedure, which is used by expand.scm, chicken-syntax.scm and
chicken-ffi-syntax.scm.

The "cond-expand" macro has been moved to the "initial macro
environment", so that it's always available, even inside modules. The
macro probably really belongs in (chicken base) analogously to where it
lives in r7rs (on the other hand, it also is available as a core subform
in the "library" form), but having it initially available makes it much
easier to have code that works both in CHICKEN 4 and CHICKEN 5 by using
cond-expand to determine which imports to use; for the majority of
programs that'll be the only substantial difference between CHICKEN 4
and 5.

Tests have been updated to more correctly import only the stuff they
truly need from scheme, chicken.base or chicken.module, as the bare
"chicken" module is on its way out.

Signed-off-by: Evan Hanson <evhan@foldling.org>