Fix output file extension for `csc -c foo.c`
This expands the changes in 3eaa4f4 to include the situation where the
input file is a C source file. It also does some minor refactoring
around the target filename handling in csc.scm, adding a helper
procedure and changing [] to (), and adds a test for csc's handling of
the "-c" and "-t" flags.