Clean up all generated files when running "make testclean"
Update the test scripts and "testclean" make target to remove all files
generated by the tests.

Add an ".out" suffix to generated files so they can be easily
distinguished.

Escape the ">" in "string->number" when calling echo in runtests.bat.
This avoids accidentally creating a file called "number" when running
the test suite on Windows.

Ignore the exit status of the USES_SONAME library removal command in the
"clean" target.

Always invoke diff(1) with expected on the left, actual on the right.

Signed-off-by: Peter Bex <peter@more-magic.net>