Add module hack so library.scm can refer to built-in core macros
This is necessary because library.scm is defining some modules, from
which it also uses macros. For "scheme" this is possible because it is
built-in as a special case but for srfi-12/chicken.condition and
chicken.base we can't import them from those libraries themselves.

Currently we get by through the "chicken" module which contains
secondary exports for all these macros because they used to be available
from that module, but the "chicken" module will disappear soonish.

This commit is in preparation for the drop of the "chicken" module,
which can then be done once we make a prerelease snapshot containing
this change  We can then add the following import to library.scm:
(import (chicken internal syntax))

Signed-off-by: Evan Hanson <evhan@foldling.org>