Drop `read-string' rewrite from c-platform.scm
After 25064a3c, the c-platform rewrite for `read-string' would ignore
its second argument and use an internal version of the procedure
whenever at least one argument was given, leading to potentially unsafe
behaviour since the type of that second argument wouldn't be checked (it
must be an *open* input port).

To fix this, we simply remove the rewrite so the port argument is always
checked at runtime.

Signed-off-by: Peter Bex <peter@more-magic.net>