Similar to 118fce and 596332, clean up old state upon fork
When fork receives a thunk to run in the child process, call it with a
clean, empty continuation rather than keeping any cruft from the
parent around.

This should allow GC of objects that are no longer necessary and
improve memory use somewhat, and perhaps prevent situations like
that in #1367 when using processes instead of threads.

Signed-off-by: felix <felix@call-with-current-continuation.org>