Avoid expanding to unbound procedures in FFI macros.
These macros call procedures that are defined inside the compiler
only.  This will result in an "unbound variable" error in DEBUGBUILD
CHICKENs, and a segfault(!) in non-DEBUGBUILDs.

So, we check to see if the macro expander is running inside the
compiler and emit an error if that's not the case.

Signed-off-by: felix <felix@call-with-current-continuation.org>