Make and-let* behave as specified in SRFI-2
When the body is missing the construction should return the last
expression that's been evaluated or the last variable that's been bound.
If there are no bindings and no body then #t is returned, like (and)
does.
Make sure the non-braced expressions refer to a variable and throw a
syntax error otherwise.

Signed-off-by: Peter Bex <peter@more-magic.net>
Signed-off-by: felix <felix@call-with-current-continuation.org>