Move several global identifiers into suitable modules.
We move all "exit"-related stuff to (chicken base), force-finalizers
to (chicken gc) and return-to-host to (chicken platform).

So, we move:

- exit
- emergency-exit
- on-exit
- exit-handler
- implicit-exit-handler
- force-finalizers
- return-to-host

The reason the "exit" stuff moves to (chicken base) and not
(chicken process-context) is because the latter lives in the posix
unit, which means we'd have to link in a large unit in programs that
just want to exit, which is absurd.

Also, exit is not really process _context_ as such, and the cleanup
stuff is so "core" that it doesn't belong in posix.

Signed-off-by: Evan Hanson <evhan@foldling.org>