Make default imports consistent between interpreter and compiler
This moves default-(syntax-)imports are moved from c-platform.scm into
internal.scm so that it can be found by csi.scm as well.

The imports are changed to scheme, chicken.base and chicken.syntax
in order to prepare for eventual removal of the "chicken" module.
Tests are changed to import the needed modules now that "chicken"
isn't in the default modules anymore.

Signed-off-by: felix <felix@call-with-current-continuation.org>