Generalise result type when scrutiniser merges differently-valued procedures
When merging procedures with different result counts, we have to fall
back to an "any" result since there's currently no way to express that a
procedure may have, for example, zero or one results.

We also generalise union types that include a "noreturn" result to the
"any" type, since we can't (currently) do anything useful with
procedures that are potentially-but-not-certainly noreturn.

Fixes #1399.

Signed-off-by: Peter Bex <peter@more-magic.net>