Fix calls to C_reclaim to use the number of saved objects (#1428)
When the argvector change was introduced, several C_reclaim calls were
incorrectly changed to use the size of the argvector to the C function
that would invoke C_reclaim.  In many cases we don't save all the
arguments, which means "c" would be larger than the size of the
temporary stack.

The value of "c" gets set to C_restart_c, which will be used in the
setjmp trampoline as the size of the buffer to set aside for the
re-invoked closure's argvector.  Then, we read C_restart_c words from
the stack top, which might overshoot the temporary stack bottom, which
means we're potentially reading beyond malloced memory.

Finally, CHICKEN_initialize would save 2 objects on the temp stack but
neglected to set C_restart_c to the correct value.  This appeared to
work only because CHICKEN_run happened to look at the effective stack
size rather than the restart count.  Now, we use the restart count and
also assert that it's always identical to the stack size.  Any
mismatch must be due to a bug.

This is not an exploitable bug because the copied bytes are
inaccessible: they're lying just beyond the highest argument in the
argvector, which is inaccessible from Scheme.  The worst that can
happen is that we try to read from a page that doesn't belong to us
and we'll segfault.

Signed-off-by: felix <felix@call-with-current-continuation.org>