Remove obsolete/deprecated stuff
This cleans up the mess of variables from "scheme" and "chicken.base"
that were aliased at toplevel, as there should be nothing relying on
that anymore.

Also add some explicit (import scheme) to non-module files, because
we're using -explicit-use to compile most files, this is required
even though these are not inside a module.

Finally, remove support for direct lambdas in define-syntax or
let-syntax, in favour of requiring wrapped transformer objects as
created by [ei]r-macro-transformer and syntax-rules.

Signed-off-by: felix <felix@call-with-current-continuation.org>