Fix MSYS tests by using Windows-style path separators in runtests.sh
When executing the runtests.sh script on Windows (in other words, when
testing on MSYS), we need to use ";" as the path component separator for
the CHICKEN_REPOSITORY_PATH variable. We use the presence of the MSYSTEM
environment variable to detect this situation, as is done when setting
the TEST_DIR variable.