Split process-context library into "standard" and "posix" components
This moves the high-level procedures of the chicken.process-context
module out of posix.scm and into library.scm, to avoid programs having
to load the posix unit for things like `get-environment-variable'. The
procedures that are left in the "posix" component are those that deal
with file descriptors, user/group information, and chroot(2).

To allow the setter for `current-directory' to continue to be used with
file descriptor arguments, a hook for directory changes is introduced.
By default, this hook invokes the standard `change-directory' procedure,
and when the posix unit is loaded it's extended to support fchdir(2).

Also, evict the procedures in chicken.process-context from the bare
"chicken" module, in preparation for its removal.

Signed-off-by: Peter Bex <peter@more-magic.net>