Disable finalizer-test for toplevel literals
This test fails on a regular basis for not directly related reasons,
depending on toplevel allocation patterns and GC-frequency. Since
the situation is relatively obscure (finailzer on code literal in evaluated
code only) and since faiilures are frequently triggered by changes
that do not directly refer to this, I have temporarily disabled the
test until we settle on a desired semantics or a complete and final
fix.