Make the "eval-modules" unit depend on "modules"
The "eval-modules" unit needs to load the "modules" unit, since the
`##sys#import-library-hook' procedure must be defined before it can be
redefined in eval-modules.scm. So, we add a dependency on "eval" (since
"eval" loads "modules", but must itself be loaded first).

Also, drop the "eval" dependency from "modules", since "modules" must be
loaded before "eval", and add a comment explaining why. This is similar
to the relationship between "expand" and "modules", which already has a
comment to explain it.

Make "modules" use the "chicken-syntax" unit, since "modules" needs the
various ##sys#-prefixed macro environments to be defined before it can
refer to them when registering core modules. "chicken-syntax" will then
load "expand" and "internal" as necessary.