Move port procedures out of toplevel and drop most "chicken" imports
Move the following procedures into the "chicken.base" library:

  port? port-closed? input-port-open? output-port-open?
  get-output-string open-input-string open-output-string
  flush-output

Move the following procedures into the "chicken.port" library:

  port-name port-position set-port-name!

Go through all import expressions and remove "chicken", where possible.
Where not possible, move the "chicken" import into its own expression
with all of the identifiers that it's still necessary to import clearly
indicated.

Import "chicken.syntax" in chicken-syntax.scm and use the imported
identifier names directly, rather than manually including the module
prefix. This is possible now that we're bootstrapping with a version of
CHICKEN that itself includes a "chicken.syntax" module.

Fix the c-platform rewrite for `port?', which was not actually in the
(scheme) module and now lives in "chicken.base" anyway.

Signed-off-by: Peter Bex <peter@more-magic.net>