Drop renaming code that was never possibly being used
This was changed a while ago, and it turns out that the "fix" was
actually disabling some code because assq will return a pair, not a
symbol, so the symbol? check would always fail.

The code in the else branch that would always trigger is duplicated in
the main cond's else branch, so we can drop the entire assq cond
clause.

Signed-off-by: Evan Hanson <evhan@foldling.org>