Revert eval-modules changes so unit isn't included in dynamic libchicken
This fixes commits 4f39a4d0 and d35a74ce, which erroneously included the
eval-modules unit in the dynamic library (i.e. libchicken.so) as well as
the static one (libchicken.a).

The eval-modules-static.o artifact *should* be included in the list of
LIBCHICKEN_STATIC_OBJECTS, but not in LIBCHICKEN_SCHEME_OBJECTS_1 since
that will also put it in LIBCHICKEN_SHARED_OBJECTS where it doesn't
belong (eval-modules is for static compilation only).

Signed-off-by: felix <felix@call-with-current-continuation.org>