Track all potential values for use when generating assigned global lambda infos
This changes the "potential-value" item in the analysis database from a
single value to a list of values, so that all assignments to a global
can be tracked and used when generating lambda info. Previously, only
the last assignment would be remembered, with any previously-encountered
potential value being clobbered.

Also, print a symbol's potential values even when a variable's true
value is unknown when printing the analysis database, and print them
last (after "refs", "css", and so on).

Fixes #1363.

Signed-off-by: felix <felix@call-with-current-continuation.org>