Speed up compiled module registration by avoid unnecessary work
The merge-se calls are still of quadratic complexity, but we can avoid
doing a lot of the work when there are no macros exported by a module,
since we do not need to patch up any syntax environments.

Also avoid consing up a full list if all syntax environments are
empty; we can just use that one list.

Signed-off-by: felix <felix@call-with-current-continuation.org>