Check -L and -I for valid <DIR>
In csc, you can pass -L<DIR> and -I<DIR> which will get passed on to
the C compiler with the -L prefix, like this:

    csc -L/tmp/lib

This is handy. However, -L and -I are overloaded to pass options to
the C compiler that do not include the prefix, like this:

    csc -L -lpng

In the first example, the compiler sees "-L/tmp/lib" and in the
seconds it sees only "-lpng". This causes problems if second version
is accidentally quoted, like this:

    csc "-L -lpng"

Now C compiler will see "-L -lpng" which isn't an error but it should
be. We rarely have library directories that are called " -lpng". This
patch addresses that. If you actually have a library path that starts
with space, you will have to do something like `csc "-L./ madness"`.

Signed-off-by: Kooda <kooda@upyum.com>