Simplify eval/meta a bit and drop ##sys#active-eval-environment
In practice, the active-eval-environment will always be set to the
parameter ##sys#current-environment (_not_ its value!), and there
is no way to override it (anymore?).

Instead of manually doing a dynamic-wind for all the parameters and
manually set/restore it to the correct values (which is error-prone)
we'll just use parameterize.  Only the meta environments will be
manually reset to the current value of their corresponding
environments one level down in the phasing tower (which is the only
depth currently supported by CHICKEN).

If we ever implement infinite phasing support, I think we can change
this to use a stack, instead.

Signed-off-by: Kooda <kooda@upyum.com>